Fix the limit clause applying on the SegmentGeneral/General locus
When creating a limit path on the top of a subpath, which is
unorder and with SegmentGeneral/General locus, the locus type
of the Limit path should be no longer SegmentGeneral/General.
Because it may produce different output on different segments.
For example:
```
gpadmin=# explain insert into rpt select i from rpt2 limit 1;
                                  QUERY PLAN
------------------------------------------------------------------------------
 Insert on rpt  (cost=0.00..431.03 rows=1 width=4)
   ->  Result  (cost=0.00..431.00 rows=1 width=8)
         ->  Limit  (cost=0.00..431.00 rows=1 width=4)
               ->  Seq Scan on rpt rpt2  (cost=0.00..431.00 rows=10 width=4)
```
The table `rpt` and `rpt2` are replicated tables. The subquery
`select i from rpt2 limit 1;` may be different on different segments,
which will lead the replicated table to an inconsistent state.

With this patch, the subquery is executed only on one segment, and
the result is broadcasted to all segments. The plan is:
```
gpadmin=# explain insert into rpt select i from rpt2 limit 1;
                                        QUERY PLAN
-------------------------------------------------------------------------------------------
 Insert on rpt  (cost=1063.00..1063.05 rows=1 width=4)
   ->  Broadcast Motion 1:3  (slice1; segments: 1)  (cost=1063.00..1063.05 rows=3 width=4)
         ->  Limit  (cost=1063.00..1063.00 rows=1 width=4)
               ->  Seq Scan on rpt2  (cost=0.00..1063.00 rows=96300 width=4)
```

Co-authored-by: Zhenghua Lyu zlv@pivotal.io