Change default gp_role if command line doesn't have option
Fix pg_ctl when hotstandby is turned on,

When the standby turns on hotstandby, the final postmaster status
of standby node becomes 'ready', not 'standby'.

Currently, the coordinator has no replication slot for the standby.
The standby priority in walsender is always 0.
With pg_auto_failover, it creates a replication slot, so the priority
in walsender may not be 0.