gpconfig: add confirmation to continue for user when a host is down (#11764)
* gpconfig: add confirmation when a host is down
When a host is down, and a user runs gpconfig to update any config: we
should warn them and ask for confirmation to continue with the config
update. Since they might get out of sync config values.

The warning and ask for confirmation were part of gpconfig until 6X,
which got removed during py2/3 migration change.

We now won't try to update the unreachable hosts if the user says
continue, and only update config on reachable hosts. This change would
make gpconfig not show any further warnings/errors about those
unreachble hosts

* behave: refactor gpconfig behave tests steps
We need to verify whether we did not change the config file.  This can
occur when a host is unreachable and the user chooses to abort the
config update.

Co-authored-by: Kalen Krempely <kkrempely@vmware.com>